requirements.txt
# Pachete de bază
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
opencv-python-headless==4.8.0.74
pillow==10.0.0
gradio==4.0.0
scikit-image==0.21.0
requests==2.31.0
tqdm==4.65.0
psutil==5.9.5

# Diffusers și transformers
diffusers==0.25.0
transformers==4.36.0
accelerate==0.25.0
safetensors==0.4.0
huggingface_hub==0.19.0

# Modele de segmentare
mediapipe==0.10.5
segment-anything==1.0
ultralytics==8.0.20

# Unelte specializate
rembg==2.0.50
xformers==0.0.22
bitsandbytes==0.41.1